% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{verifier}
\alias{verifier}
\title{Check if a colour is valid}
\usage{
verifier(x)
}
\arguments{
\item{x}{Colour string}
}
\value{
boolean for successful conversion or not
}
\description{
Check if a colour is valid
}
\examples{
dummyimage:::verifier("white")
dummyimage:::verifier("///")
}
